refactor(apps): Phase 2 — migrate visitors, departments, users to Zod codecs
First bidirectional converters on the codec pattern. Behaviour unchanged —
the Phase 0 golden snapshots (both directions) still pass.

- Split the mapped-codec helper into reusable `mappedDecode` / `mappedEncode`
  transforms so a codec can pair a mapped decode with a bespoke encode.
- `VisitorCodec` / `DepartmentCodec`: mapped decode (bucketed) + bespoke
  encode matching `convertAppVisitor` (status default, truthy-conditional
  fields, no `_updatedAt`/`activity` write-back) and `convertAppDepartment`
  (unconditional inverse), each merging `_unmappedProperties_`.
- `UserCodec`: bespoke both ways via the shared enum codecs; the contextual
  invalid-status warning stays in the codec (it needs the user id/username),
  and the pre-existing `utfOffset` typo is preserved verbatim.
- Delegate the three converter classes to their codecs.
