+++ apps/meteor/app/apps/server/converters/codecs/index.ts
@@ -1,6 +1,9 @@
 export { RoomTypeCodec, SettingTypeCodec, UserStatusConnectionCodec, UserTypeCodec } from './enums';
-export { createMappedCodec } from './mappedData';
+export { createMappedCodec, mappedDecode, mappedEncode } from './mappedData';
 export type { FieldMap } from './mappedData';
 export { SettingCodec } from './settings';
 export { RoleCodec } from './roles';
 export { VideoConferenceCodec } from './videoConferences';
+export { VisitorCodec } from './visitors';
+export { DepartmentCodec } from './departments';
+export { UserCodec } from './users';
